moved ClassSelector and ControlPanel components to home folder
