+++ src/components/home/ControlPanel.js
